[share] MethodCallHandler.java uses unchecked or unsafe operations
