--- packages/share/android/src/main/java/io/flutter/plugins/share/ShareFilesMap.java
@@ -1,45 +1,0 @@
-package io.flutter.plugins.share;
-
-import java.util.List;
-
-public class ShareFilesMap {
-    List<String> paths;
-    List<String> mimeTypes;
-    String text;
-    String subject;
-
-    public ShareFilesMap() {
-    }
-
-    public List<String> getPaths() {
-        return paths;
-    }
-
-    public void setPaths(List<String> paths) {
-        this.paths = paths;
-    }
-
-    public List<String> getMimeTypes() {
-        return mimeTypes;
-    }
-
-    public void setMimeTypes(List<String> mimeTypes) {
-        this.mimeTypes = mimeTypes;
-    }
-
-    public String getText() {
-        return text;
-    }
-
-    public void setText(String text) {
-        this.text = text;
-    }
-
-    public String getSubject() {
-        return subject;
-    }
-
-    public void setSubject(String subject) {
-        this.subject = subject;
-    }
-}
